+++ scripts/components/profiles/map.component.js
@@ -3,6 +3,7 @@ import { json as d3_json } from 'd3-request';
 import { geoPath as d3_geoPath, geoMercator as d3_geoMercator } from 'd3-geo';
 import { geoRobinson as d3_geoRobinson } from 'd3-geo-projection';
 import * as topojson from 'topojson';
+import choroLegend from 'components/profiles/choro-legend.component';
 
 function getFeaturesBox(featureBounds) {
   return {
@@ -22,10 +23,9 @@ function fitGeoInside(featureBounds, width, height) {
   return { scale, trans };
 }
 
-
-export default (className, { topoJSONPath, topoJSONRoot, getPolygonClassName, showTooltipCallback, hideTooltipCallback, useRobinsonProjection }) => {
-
+export default (className, { topoJSONPath, topoJSONRoot, getPolygonClassName, showTooltipCallback, hideTooltipCallback, useRobinsonProjection, legend }) => {
   const d3Container =  d3_select(className);
+  d3Container.node().classList.remove('-with-legend');
   const containerComputedStyle = window.getComputedStyle(d3Container.node());
   const width = parseInt(containerComputedStyle.width);
   const height = parseInt(containerComputedStyle.height);
@@ -77,4 +77,10 @@ export default (className, { topoJSONPath, topoJSONRoot, getPolygonClassName, sh
 
     container.selectAll('path').style('stroke-width', .5 / scale);
   });
+
+  if (legend) {
+    d3Container.append('div').attr('class', 'legend-container');
+    choroLegend(className, '.legend-container', legend);
+  }
+
 };