+++ scripts/containers/tool/map-dimensions.container.js
@@ -1,27 +1,26 @@
 import connect from 'connect';
-import { selectMapDimension } from 'actions/tool.actions';
-import { toggleModal, loadTooltip } from 'actions/app.actions';
 import MapDimensions from 'components/tool/map-dimensions.component.js';
+import { toggleMapSidebarGroup, toggleMapDimension } from 'actions/tool.actions';
+import { loadTooltip } from 'actions/app.actions';
 
 const mapMethodsToState = (state) => ({
   loadMapDimensions: {
-    _comparedValue: (state) => state.tool.mapDimensions,
+    _comparedValue: (state) => state.tool.mapDimensionsGroups,
     _returnedValue: (state) => {
-      return state.tool.mapDimensionsFolders.map(dimensionGroup => {
-        return {
-          dimensionGroup,
-          dimensions: state.tool.mapDimensions.filter(dimension => dimension.groupId === dimensionGroup.id)
-        };
-      });
+      return {
+        mapDimensionsGroups: state.tool.mapDimensionsGroups,
+        expandedMapSidebarGroupsIds: state.tool.expandedMapSidebarGroupsIds
+      };
     }
   },
   selectMapDimensions: state.tool.selectedMapDimensions,
+  toggleSidebarGroups: state.tool.expandedMapSidebarGroupsIds
 });
 
 const mapViewCallbacksToActions = () => ({
-  onMapDimensionsSelected: dimensionData => selectMapDimension(dimensionData),
-  onToggleModal: (visibility, data) => toggleModal(visibility, data),
   onMapDimensionsLoaded: () => loadTooltip(),
+  onToggleGroup: (id) => toggleMapSidebarGroup(id),
+  onDimensionClick: (uid) => toggleMapDimension(uid)
 });
 
 export default connect(MapDimensions, mapMethodsToState, mapViewCallbacksToActions);