+++ scripts/pages/profile-actor.page.js
@@ -27,11 +27,15 @@ import formatApostrophe from 'utils/formatApostrophe';
 import formatNumber from 'utils/formatNumber';
 import _ from 'lodash';
 import { getURLFromParams, GET_ACTOR_FACTSHEET } from '../utils/getURLFromParams';
+import { ACTORS_TOP_SOURCES_SWITCHERS_BLACKLIST } from 'constants';
+import TopSourceTemplate from 'ejs!templates/profiles/top-source-switcher.ejs';
 
 const defaults = {
   country: 'Brazil',
   commodity: 'soy'
 };
+
+const tooltip = new Tooltip('.js-infowindow');
 
 const _onSelect = function(value) {
   this.setTitle(value);
@@ -39,11 +43,11 @@ const _onSelect = function(value) {
 };
 
 const _build = (data, nodeId) => {
-  const tooltip = new Tooltip('.js-infowindow');
   const lineSettings = {
-    margin: { top: 10, right: 100, bottom: 25, left: 94 },
+    margin: { top: 10, right: 100, bottom: 30, left: 94 },
     height: 244,
     ticks: {
+      xTicks: 6,
       yTicks: 6,
       yTickPadding: 10,
       yTickFormatType: 'top-location',
@@ -65,16 +69,15 @@ const _build = (data, nodeId) => {
     }
   };
 
-
   if (data.top_sources.municipality.lines.length) {
-    document.querySelector('.js-top-municipalities-title').innerHTML = `Top source regions of ${formatApostrophe(_.capitalize(data.node_name))} soy: municipalities`;
-    let topMunicipalitiesLines = data.top_sources.municipality;
+    _setTopSourceSwitcher(data);
+    const topMunicipalitiesLines = Object.assign({}, data.top_sources.municipality);
     topMunicipalitiesLines.lines = topMunicipalitiesLines.lines.slice(0, 5);
     new Line(
       '.js-top-municipalities',
       topMunicipalitiesLines,
       data.top_sources.included_years,
-      lineSettings,
+      Object.assign({}, lineSettings, { margin: { top: 10, right: 100, bottom: 25, left: 37 } }),
     );
 
     Map('.js-top-municipalities-map', {
@@ -83,8 +86,8 @@ const _build = (data, nodeId) => {
       getPolygonClassName: ({ properties }) => {
         const municipality = data.top_sources.municipality.lines
           .find(m => (properties.geoid === m.geo_id));
-        let value = 0;
-        if (municipality) value = municipality.value9 || 0;
+        let value = 'n-a';
+        if (municipality) value = municipality.value9 || 'n-a';
         return `-outline ch-${value}`;
       },
       showTooltipCallback: ({ properties }, x, y) => {
@@ -105,13 +108,17 @@ const _build = (data, nodeId) => {
       },
       hideTooltipCallback: () => {
         tooltip.hideTooltip();
+      },
+      legend: {
+        title: ['Soy exported in 2015', '(t)'],
+        bucket: [data.top_countries.buckets[0], ...data.top_countries.buckets]
       }
     });
   }
 
   if (data.top_countries.lines.length) {
     document.querySelector('.js-top-map-title').innerHTML = `Top destination countries of ${formatApostrophe(_.capitalize(data.node_name))} soy`;
-    let topCountriesLines = data.top_countries;
+    const topCountriesLines = Object.assign({}, data.top_countries);
     topCountriesLines.lines = topCountriesLines.lines.slice(0, 5);
     new Line(
       '.js-top-destination',
@@ -127,8 +134,8 @@ const _build = (data, nodeId) => {
       getPolygonClassName: ({ properties }) => {
         const country = data.top_countries.lines
           .find(c => (properties.name.toUpperCase() === c.name.toUpperCase()));
-        let value = 0;
-        if (country) value = country.value9 || 0;
+        let value = 'n-a';
+        if (country) value = country.value9 || 'n-a';
         return `-outline ch-${value}`;
       },
       showTooltipCallback: ({ properties }, x, y) => {
@@ -149,16 +156,19 @@ const _build = (data, nodeId) => {
       },
       hideTooltipCallback: () => {
         tooltip.hideTooltip();
+      },
+      legend: {
+        title: ['Soy exported in 2015', '(t)'],
+        bucket: [data.top_countries.buckets[0], ...data.top_countries.buckets]
       }
     });
-
   }
 
   if (data.sustainability.length) {
     new MultiTable({
       el: document.querySelector('.js-sustainability-table'),
       data: data.sustainability,
-      tabsTitle: `Sustainability of ${formatApostrophe(data.node_name)} TOP source regions in 2015:`,
+      tabsTitle: `Sustainability of ${formatApostrophe(data.node_name)} top source regions in 2015:`,
       type: 't_head_actors',
       target: 'actor'
     });
@@ -167,7 +177,7 @@ const _build = (data, nodeId) => {
   new Scatterplot('.js-companies-exporting', {
     data: data.companies_exporting.companies,
     xDimension: data.companies_exporting.dimensions_x,
-    nodeId: nodeId,
+    node: { id: nodeId, name: data.node_name },
     showTooltipCallback: (company, indicator, x, y) => {
       tooltip.showTooltip(x, y, {
         title: company.name,
@@ -192,10 +202,10 @@ const _build = (data, nodeId) => {
 };
 
 const _setInfo = (info, nodeId) => {
-  document.querySelector('.js-name').innerHTML = info.name ? _.capitalize(info.name) : '-';
+  document.querySelector('.js-name').textContent = info.name ? _.capitalize(info.name) : '-';
   document.querySelector('.js-link-button-name').textContent = formatApostrophe(_.capitalize(info.name)) + ' PROFILE';
-  document.querySelector('.js-legend').innerHTML = info.type || '-';
-  document.querySelector('.js-country').innerHTML = info.country ? _.capitalize(info.country) : '-';
+  document.querySelector('.js-legend').textContent = info.type || '-';
+  document.querySelector('.js-country').textContent = info.country ? _.capitalize(info.country) : '-';
   if (info.forest_500 > 0) document.querySelector('.js-forest-500-score .circle-icon[data-value="1"] use').setAttributeNS('http://www.w3.org/1999/xlink', 'href', '#icon-circle-filled');
   if (info.forest_500 > 1) document.querySelector('.js-forest-500-score .circle-icon[data-value="2"] use').setAttributeNS('http://www.w3.org/1999/xlink', 'href', '#icon-circle-filled');
   if (info.forest_500 > 2) document.querySelector('.js-forest-500-score .circle-icon[data-value="3"] use').setAttributeNS('http://www.w3.org/1999/xlink', 'href', '#icon-circle-filled');
@@ -208,7 +218,7 @@ const _setInfo = (info, nodeId) => {
   }
   document.querySelector('.js-link-map').setAttribute('href', `./flows.html?selectedNodesIds=[${nodeId}]&isMapVisible=true`);
   document.querySelector('.js-link-supply-chain').setAttribute('href', `./flows.html?selectedNodesIds=[${nodeId}]`);
-  document.querySelector('.js-summary-text').innerHTML = info.summary ? info.summary : '-';
+  document.querySelector('.js-summary-text').textContent = info.summary ? info.summary : '-';
 };
 
 const _setEventListeners = () => {
@@ -223,13 +233,110 @@ const _showErrorMessage = () => {
   el.querySelector('.js-error-message').classList.remove('is-hidden');
 };
 
+const _setTopSourceSwitcher = (data) => {
+  const template = TopSourceTemplate({
+    nodeName: formatApostrophe(_.capitalize(data.node_name)),
+    switchers: Object.keys(data.top_sources).filter(key => !(ACTORS_TOP_SOURCES_SWITCHERS_BLACKLIST.includes(key)))
+  });
+  document.querySelector('.js-top-municipalities-title').innerHTML = template;
+
+  const switchers = Array.prototype.slice.call(document.querySelectorAll('.js-top-source-switcher'), 0);
+  switchers.forEach(switcher => {
+    switcher.addEventListener('click', (e) => _switchTopSource(e, data));
+  });
+};
+
+const _switchTopSource = (e, data) => {
+  const selectedSwitch = e && e.currentTarget;
+  if (!selectedSwitch) {
+    return;
+  }
+
+  const selectedSource = selectedSwitch.getAttribute('data-key');
+  const switchers = Array.prototype.slice.call(document.querySelectorAll('.js-top-source-switcher'), 0);
+  switchers.forEach(switcher => {
+    switcher.classList.remove('selected');
+  });
+  selectedSwitch.classList.add('selected');
+
+  const topMunicipalitiesLines = Object.assign({}, data.top_sources[selectedSource]);
+  topMunicipalitiesLines.lines = topMunicipalitiesLines.lines.slice(0, 5);
+  new Line(
+    '.js-top-municipalities',
+    topMunicipalitiesLines,
+    data.top_sources.included_years,
+    {
+      margin: { top: 10, right: 100, bottom: 25, left: 37 },
+      height: 244,
+      ticks: {
+        xTicks: 6,
+        yTicks: 6,
+        yTickPadding: 10,
+        yTickFormatType: 'top-location',
+        xTickPadding: 15
+      },
+      showTooltipCallback: (location, x, y) => {
+        tooltip.showTooltip(x, y, {
+          title: `${data.node_name} > ${location.name.toUpperCase()}, ${location.date.getFullYear()}`,
+          values: [
+            {
+              title: 'Trade Volume',
+              value: formatNumber(location.value),
+              unit: 'Tons'
+            }
+          ]
+        });
+      },
+      hideTooltipCallback: () => {
+        tooltip.hideTooltip();
+      }
+    },
+  );
+
+  document.querySelector('.js-top-municipalities-map').innerHTML = '';
+  Map('.js-top-municipalities-map', {
+    topoJSONPath: `./vector_layers/${defaults.country.toUpperCase()}_${selectedSource.toUpperCase()}.topo.json`,
+    topoJSONRoot: `${defaults.country.toUpperCase()}_${selectedSource.toUpperCase()}`,
+    getPolygonClassName: ({ properties }) => {
+      const source = data.top_sources[selectedSource].lines
+        .find(s => (properties.geoid === s.geo_id));
+      let value = 'n-a';
+      if (source) value = source.value9 || 'n-a';
+      return `-outline ch-${value}`;
+    },
+    showTooltipCallback: ({ properties }, x, y) => {
+      const source = data.top_sources[selectedSource].lines
+        .find(s => (properties.geoid === s.geo_id));
+      let title = `${data.node_name} > ${properties.nome.toUpperCase()}`;
+      let body = null;
+      if (source) body = source.values[0];
+
+      tooltip.showTooltip(x, y, {
+        title,
+        values: [{
+          title: 'Trade Volume',
+          value: formatNumber(body),
+          unit: 'Tons'
+        }]
+      });
+    },
+    hideTooltipCallback: () => {
+      tooltip.hideTooltip();
+    },
+    legend: {
+      title: ['Soy exported in 2015', '(t)'],
+      bucket: [data.top_sources.buckets[0], ...data.top_sources.buckets]
+    }
+  });
+};
+
 const _init = ()  => {
   const url = window.location.search;
   const urlParams = getURLParams(url);
   const nodeId = urlParams.nodeId;
   const commodity = urlParams.commodity || defaults.commodity;
 
-  const actorFactsheetURL = getURLFromParams(GET_ACTOR_FACTSHEET, { node_id: nodeId }, true);
+  const actorFactsheetURL = getURLFromParams(GET_ACTOR_FACTSHEET, { node_id: nodeId });
 
   fetch(actorFactsheetURL)
     .then((response) => {