+++ CHANGELOG.md
@@ -1,4 +1,16 @@
+
 - tool: fixed various issues with the map with contexts that don't have subnational granularity
+- profiles: fixes order of lines in Line chart
+- profiles: improves xAxis ticks on Line chart (display months in number of years >2)
+- profiles: fixes loader showing up on top of nav.
+- profiles: fixes choropleth maps only showing 5 values.
+- profiles: fixes scatterplot not showing current node.
+- profiles: fixes municipalities legend not showing when switching source.
+- profiles: includes map choropleth legend
+- profiles: includes the top sources' source switch.
+- tool: map dimensions groups are now collapsible
+- tool: map dimensions selecting is simpler (singl radio buttons column)
+- tool: fixed tool crash when links API fails
 - profiles: connects choropleth maps to real data in actors profile pages
 - home: only allow video playing on full screen
 - home: show screenshot instead of video when video playback has not started yet
@@ -6,6 +18,7 @@
 - tool: fixed water scarcity color scale
 - tool: added link tooltips resize by information
 - tool: added link tooltips quant units
+- profiles: actor municipalities/biome/state switcher
 - profiles: added line chart locations on actors page
 - profiles: added hover effects on shortcut buttons
 - profiles: companies exporting chart (mock)